fix: middleware import issues related to url stringify
