+++ apps/hobbies-helsinki/src/middleware.ts
@@ -1,11 +1,9 @@
+// import { DEFAULT_LANGUAGE } from 'events-helsinki-components';
 import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
-
-import { stringifyUrlObject } from './utils/routerUtils';
+import stringifyUrlObject from 'utils/stringifyUrlObject';
 
 // TODO: For some reason middleware cannot read `'events-helsinki-components` package without breaking the build
-// import { DEFAULT_LANGUAGE } from 'events-helsinki-components';
-
 const DEFAULT_LANGUAGE = 'fi';
 
 const requestType = {
